requirements.txt
asgiref==3.4.1
black==21.9b0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.2
dj-database-url==0.5.0
Django==3.2.8
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
idna==3.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dotenv==0.19.1
pytz==2021.3
redis==3.5.3
regex==2021.10.8
requests==2.26.0
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
